# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Inkwell API
  version: "0.2"
  description: |
    API-first Markdown publishing platform.

    All write endpoints require `X-Api-Key`. Reads are public unless stated
    otherwise. See [docs/API.md](./API.md) for the full narrative reference.
  contact:
    name: Inkwell
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your-railway-domain.up.railway.app
    description: Production (replace with actual domain)
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Shared admin key (`INKWELL_API_KEY`) or a scoped token
        (`ink_<prefix>_<secret>`).
  schemas:
    Document:
      type: object
      required:
        - id
        - slug
        - title
        - bodyMarkdown
        - renderedHtml
        - status
        - growth
        - tags
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        slug:
          type: string
          example: "hello-world"
          maxLength: 200
        title:
          type: string
          example: "Hello World"
          maxLength: 500
        bodyMarkdown:
          type: string
          example: "# Hello World\n\nBody text here."
          maxLength: 262144
        renderedHtml:
          type: string
          example: "<h1>Hello World</h1>\n<p>Body text here.</p>"
        status:
          type: string
          enum: [draft, published]
        growth:
          type: string
          enum: [seedling, budding, evergreen]
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
          example: ["rust", "notes"]
        version:
          type: integer
          format: int64
          description: Monotonic revision counter. Echo as `If-Match` on conditional updates.
          example: 3
        createdAt:
          type: string
          format: date-time
          example: "2026-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-06-01T08:00:00Z"
...
    UpdateDocumentRequest:
      type: object
      description: At least one field is required.
      properties:
        title:
          type: string
          maxLength: 500
        bodyMarkdown:
          type: string
          maxLength: 262144
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 20
        growth:
          type: string
          enum: [seedling, budding, evergreen]
        slug:
          type: string
          maxLength: 200
          description: Rename the document. Old slug registers as a 301 alias.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
              example: "Missing or invalid API key."
            requestId:
              type: string
              format: uuid
              description: Correlation ID. Present on every in-request error.
            slug:
              type: string
              description: Present on 409 Conflict for duplicate slug.
            allow:
              type: string
              description: Present on 405 Method Not Allowed.
              example: "GET, POST"
  responses:
    Unauthorized:
      description: Missing or invalid `X-Api-Key`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Missing or invalid API key."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    RejectedReadCredential:
      description: >
        An `X-Api-Key` was presented and rejected (revoked token, unknown prefix,
        hash mismatch, a value that is neither the shared key nor token-shaped, or
        a duplicated/non-ASCII header). Treat this as "re-authenticate", not
        "empty result" — see ADR 0015. A rejected `inkwell_session` cookie is NOT
        reported here: that channel downgrades to anonymous and returns `200` with
        published content, so a stale browser cookie never breaks public pages. A
        valid credential lacking the `read` scope also stays `200`, published only.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Missing or invalid API key."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    ServiceUnavailable:
      description: >
        The credential could not be judged because the token lookup failed against
        Postgres (ADR 0015). Distinct from `401` — the credential was never
        actually rejected — and from `500`. Retry; check `GET /readyz`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Service temporarily unavailable. Retry shortly."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    Forbidden:
      description: Valid credential, but insufficient scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "This action requires the \"write\" scope."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    NotFound:
      description: Resource does not exist (or caller cannot see it).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "No document with slug \"hello-world\"."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Rate limit exceeded. Slow down and retry later."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
paths:
...
  /documents/{slug}:
    parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string

    get:
      summary: Get a document
      operationId: getDocument
      tags: [Documents]
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Document found.
          headers:
            ETag:
              schema:
                type: string
              description: Current version as a quoted integer e.g. `"3"`. Echo as `If-Match` on updates.
            X-Request-Id:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "301":
          description: Slug was renamed. Follow `Location` to the current slug.
          headers:
            Location:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      summary: Update a document (partial)
      operationId: updateDocument
      tags: [Documents]
      security:
        - ApiKey: []
      parameters:
        - in: header
          name: If-Match
          schema:
            type: string
          description: |
            Expected version (e.g. `"3"`). Triggers a conditional update;
            returns 409 on version mismatch. Omit for an unconditional update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentRequest"
            example:
              title: "Updated Title"
              tags: ["rust"]
      responses:
        "200":
          description: Document updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Validation error or empty body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: If-Match version mismatch or duplicate slug on rename.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  message: "Document \"hello-world\" has version 4, not the expected 3. Re-read the note and retry with the current version."
                  requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    put:
      summary: Update a document (full — same semantics as PATCH)
      operationId: updateDocumentPut
      tags: [Documents]
      security:
        - ApiKey: []
      parameters:
        - in: header
          name: If-Match
          schema:
            type: string
          description: Optional optimistic-concurrency version guard.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentRequest"
      responses:
        "200":
          description: Document updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Version mismatch or duplicate slug.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      summary: Delete a document
      operationId: deleteDocument
      tags: [Documents]
      security:
        - ApiKey: []
      responses:
        "204":
          description: Document deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /documents/{slug}/preview-tokens/{prefix}:
    delete:
      summary: Revoke a preview token
      operationId: revokePreviewToken
      tags: [Preview]
      security:
        - ApiKey: []
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: prefix
          required: true
          schema:
            type: string
          example: "abc123"
      responses:
        "204":
          description: Token revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{slug}/preview:
    get:
      summary: Render a draft via a preview token (anonymous)
      operationId: previewDocument
      tags: [Preview]
      description: |
        Renders the draft document as HTML for any bearer of a valid, non-expired,
        non-revoked preview token. No `X-Api-Key` header required. Any failure
        (missing/invalid/expired/revoked token, unknown document) returns 401 — the
        draft's existence is not leaked.
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: token
          required: true
          schema:
            type: string
          example: "pvw_abc123_secretvalue"
          description: The full preview token value.
      responses:
        "200":
          description: Draft rendered as HTML.
          content:
            text/html:
              schema:
                type: string
        "401":
          description: Token missing, invalid, expired, revoked, or document not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...